Split vote verification/validation
There is functionality in Tendermint that cares about vote extensions
being present (largely confined to consensus) and functionality that
doesn't.

There will be cases where we won't have vote extensions (e.g. when
they've been pruned, as per RFC 017), and so we need to cater for vote
validation and verification in those cases.

Signed-off-by: Thane Thomson <connect@thanethomson.com>